Removal of warnings

+++ libs/qCC_db/ccCameraSensor.cpp
@@ -309,7 +309,7 @@ bool ccCameraSensor::applyViewport(ccGenericGLDisplay* win/*=0*/)
 	//aspect ratio
 	float ar = static_cast<float>(m_intrinsicParams.arrayWidth) / m_intrinsicParams.arrayHeight;
 	//fov
-	float fov_deg = m_intrinsicParams.vFOV_rad * CC_RAD_TO_DEG;
+	float fov_deg = static_cast<float>(m_intrinsicParams.vFOV_rad * CC_RAD_TO_DEG);
 	//camera position/orientation
 	ccGLMatrixd transd(trans.data());
 	win->setupProjectiveViewport(transd,fov_deg,ar);
